pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ipachecker"
dynamic = ["version"]
description = "A tool for checking if an .ipa file is encrypted"
authors = [
    {name = "Andres99"}
]
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "docopt-ng>=0.9.0",
    "rich>=14.1.0",
    "macholib>=1.16.3",
]
keywords = [
    "ios",
    "ipa",
    "analysis",
    "metadata",
    "encryption",
    "mobile",
    "app",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Archiving",
    "Topic :: Utilities",
]

[project.urls]
homepage = "https://github.com/Andres9890/ipachecker"
repository = "https://github.com/Andres9890/ipachecker.git"
issues = "https://github.com/Andres9890/ipachecker/issues"

[project.scripts]
ipachecker = "ipachecker.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "ipachecker.__version__"}

[tool.setuptools.packages.find]
include = ["ipachecker*"]